# code-template/src/main/java/com/changgou/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "微服务工程代码生成器,生成Dao、Service、Controller、Pojo、Feign、Swagger的JSON脚本",
    "version": "1.0.0",
    "title": "深圳黑马训练营代码生成器",
    "termsOfService": "https://github.com/shenkunlin/code-template.git",
    "contact": {
      "email": "itheima@itcast.cn"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "petstore.swagger.io",
  "basePath": "/v2",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/categoryReport/{id}": {
      "delete": {
        "tags": [
          "CategoryReportController"
        ],
        "summary": "根据ID删除CategoryReport",
        "description": "根据ID删除CategoryReport方法详情",
        "operationId": "deleteUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID删除CategoryReport方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID删除CategoryReport"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "put": {
        "tags": [
          "CategoryReportController"
        ],
        "summary": "根据ID修改CategoryReport",
        "description": "根据ID修改CategoryReport方法详情",
        "operationId": "updateUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改CategoryReport方法详情",
            "required": true
          },
          {
            "in": "body",
            "name": "CategoryReport",
            "description": "传入CategoryReport的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/CategoryReport"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改CategoryReport"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "get": {
        "tags": [
          "CategoryReportController"
        ],
        "summary": "根据ID查询CategoryReport",
        "description": "根据ID查询CategoryReport方法详情",
        "operationId": "findByIdUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改CategoryReport方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改CategoryReport"
            ,
            "schema": {
              "$ref": "#/definitions/Result«CategoryReport»"
            }
          }
        }
      }
    },
    "/categoryReport/search": {
      "post": {
        "tags": [
          "CategoryReportController"
        ],
        "summary": "不带分页条件搜索CategoryReport",
        "description": "不带分页条件搜索CategoryReport方法详情",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "CategoryReport",
            "description": "传入CategoryReport的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/CategoryReport"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索CategoryReport"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«CategoryReport»»"
            }
          }
        }
      }
    },
    "/categoryReport/search/{page}/{size}": {
      "get": {
        "tags": [
          "CategoryReportController"
        ],
        "summary": "分页列表查询CategoryReport",
        "description": "分页列表查询CategoryReport方法详情",
        "operationId": "findPageUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索CategoryReport"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«CategoryReport»»"
            }
          }
        }
      },
      "post": {
        "tags": [
          "CategoryReportController"
        ],
        "summary": "分页条件搜索CategoryReport",
        "description": "分页条件搜索CategoryReport方法详情",
        "operationId": "findPageUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          },
          {
            "in": "body",
            "name": "CategoryReport",
            "description": "传入CategoryReport的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/CategoryReport"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "分页条件搜索CategoryReport"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«CategoryReport»»"
            }
          }
        }
      }
    },
...
    "/order/{id}": {
      "delete": {
        "tags": [
          "OrderController"
        ],
        "summary": "根据ID删除Order",
        "description": "根据ID删除Order方法详情",
        "operationId": "deleteUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID删除Order方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID删除Order"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "put": {
        "tags": [
          "OrderController"
        ],
        "summary": "根据ID修改Order",
        "description": "根据ID修改Order方法详情",
        "operationId": "updateUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改Order方法详情",
            "required": true
          },
          {
            "in": "body",
            "name": "Order",
            "description": "传入Order的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/Order"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改Order"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "get": {
        "tags": [
          "OrderController"
        ],
        "summary": "根据ID查询Order",
        "description": "根据ID查询Order方法详情",
        "operationId": "findByIdUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改Order方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改Order"
            ,
            "schema": {
              "$ref": "#/definitions/Result«Order»"
            }
          }
        }
      }
    },
    "/order/search": {
      "post": {
        "tags": [
          "OrderController"
        ],
        "summary": "不带分页条件搜索Order",
        "description": "不带分页条件搜索Order方法详情",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "Order",
            "description": "传入Order的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/Order"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索Order"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«Order»»"
            }
          }
        }
      }
    },
    "/order/search/{page}/{size}": {
      "get": {
        "tags": [
          "OrderController"
        ],
        "summary": "分页列表查询Order",
        "description": "分页列表查询Order方法详情",
        "operationId": "findPageUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索Order"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«Order»»"
            }
          }
        }
      },
      "post": {
        "tags": [
          "OrderController"
        ],
        "summary": "分页条件搜索Order",
        "description": "分页条件搜索Order方法详情",
        "operationId": "findPageUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          },
          {
            "in": "body",
            "name": "Order",
            "description": "传入Order的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/Order"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "分页条件搜索Order"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«Order»»"
            }
          }
        }
      }
    },
...
    "/orderConfig/{id}": {
      "delete": {
        "tags": [
          "OrderConfigController"
        ],
        "summary": "根据ID删除OrderConfig",
        "description": "根据ID删除OrderConfig方法详情",
        "operationId": "deleteUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID删除OrderConfig方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID删除OrderConfig"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "put": {
        "tags": [
          "OrderConfigController"
        ],
        "summary": "根据ID修改OrderConfig",
        "description": "根据ID修改OrderConfig方法详情",
        "operationId": "updateUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改OrderConfig方法详情",
            "required": true
          },
          {
            "in": "body",
            "name": "OrderConfig",
            "description": "传入OrderConfig的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/OrderConfig"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改OrderConfig"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "get": {
        "tags": [
          "OrderConfigController"
        ],
        "summary": "根据ID查询OrderConfig",
        "description": "根据ID查询OrderConfig方法详情",
        "operationId": "findByIdUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改OrderConfig方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改OrderConfig"
            ,
            "schema": {
              "$ref": "#/definitions/Result«OrderConfig»"
            }
          }
        }
      }
    },
    "/orderConfig/search": {
      "post": {
        "tags": [
          "OrderConfigController"
        ],
        "summary": "不带分页条件搜索OrderConfig",
        "description": "不带分页条件搜索OrderConfig方法详情",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "OrderConfig",
            "description": "传入OrderConfig的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/OrderConfig"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索OrderConfig"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«OrderConfig»»"
            }
          }
        }
      }
    },
    "/orderConfig/search/{page}/{size}": {
      "get": {
        "tags": [
          "OrderConfigController"
        ],
        "summary": "分页列表查询OrderConfig",
        "description": "分页列表查询OrderConfig方法详情",
        "operationId": "findPageUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索OrderConfig"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«OrderConfig»»"
            }
          }
        }
      },
      "post": {
        "tags": [
          "OrderConfigController"
        ],
        "summary": "分页条件搜索OrderConfig",
        "description": "分页条件搜索OrderConfig方法详情",
        "operationId": "findPageUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          },
          {
            "in": "body",
            "name": "OrderConfig",
            "description": "传入OrderConfig的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/OrderConfig"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "分页条件搜索OrderConfig"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«OrderConfig»»"
            }
          }
        }
      }
    },
...
    "/orderItem/search": {
      "post": {
        "tags": [
          "OrderItemController"
        ],
        "summary": "不带分页条件搜索OrderItem",
        "description": "不带分页条件搜索OrderItem方法详情",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "OrderItem",
            "description": "传入OrderItem的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/OrderItem"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索OrderItem"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«OrderItem»»"
            }
          }
        }
      }
    },
...
    "/orderLog/search": {
      "post": {
        "tags": [
          "OrderLogController"
        ],
        "summary": "不带分页条件搜索OrderLog",
        "description": "不带分页条件搜索OrderLog方法详情",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "OrderLog",
            "description": "传入OrderLog的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/OrderLog"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索OrderLog"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«OrderLog»»"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Result": {
      "type": "object",
      "required": [
        "code",
        "data",
        "flag",
        "message"
      ],
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "description": "返回状态码,20000:成功,20001:失败,20002:用户名或密码错误,20003:权限不足,20004:远程调用失败,20005:重复操作,20006:没有对应的数据"
        },
        "data": {
          "type": "object",
          "description": "逻辑数据"
        },
        "flag": {
          "type": "boolean",
          "example": false,
          "description": "执行是否成功,true:成功,false:失败"
        },
        "message": {
          "type": "string",
          "description": "提示信息"
        }
      },
      "description": "Result"
    },
...
    "CategoryReport": {
      "type": "object",
      "properties": {
        "categoryId1": {
          "type": "integer",
          "format": "int32",
          "description": "1级分类"
        },
        "categoryId2": {
          "type": "integer",
          "format": "int32",
          "description": "2级分类"
        },
        "categoryId3": {
          "type": "integer",
          "format": "int32",
          "description": "3级分类"
        },
        "countDate": {
          "type": "date",
          "description": "统计日期"
        },
        "num": {
          "type": "integer",
          "format": "int32",
          "description": "销售数量"
        },
        "money": {
          "type": "integer",
          "format": "int32",
          "description": "销售额"
        }
      },
      "description": "CategoryReport"
    },
...
  },
...
}